requirements.txt
mysql-connector-python==2.1.6
mysqlclient==1.3.10
pandas==0.22.0
urllib3==1.22
pyspark==3.2.0

